Use `Result::into_iter` for simplification
Thanks to @Manmeetmann for teaching me this:

You can use `Result::into_iter` and then `Iterator::flatten`
to turn a `Result<T>` where `T: Iterator<Item=I>` into
an `Iterator<Item=I>` which returns no values if the original
`Result<T>` was an error.

Using this, we can do away with `IteratorMaybeEmpty` as well
as do away with the outer capture and unindent the code by
a level.